requirements.txt
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
numpy>=1.21.0
torch>=1.9.0
transformers>=4.20.0
PyPDF2>=3.0.0
python-magic>=0.4.27
flask>=2.3.0
flask-cors>=4.0.0
pyjwt>=2.8.0
werkzeug>=2.3.0
requests>=2.31.0
